// frontend/src/components/__generated__/product-detail-query.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
export type ReviewBodyFragment = {
    id: string;
    commentBody: string;
};
export type ProductReviewFragment = {
    reviews: (ReviewBodyFragment)[];
};
export type ProductDetailQuery = {
    product: ({
        id: string;
        name: string;
        description: string;
    } & ProductReviewFragment) | null;
};
export type ProductDetailQueryVariables = {
    id: string;
};