// frontend/src/components/__generated__/add-review-mutation.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
export type ReviewBodyFragment = {
    id: string;
    commentBody: string;
};
export type AddReviewMutation = {
    addReview: ({
        __typename: "Review";
    } & ReviewBodyFragment) | null;
};
export type AddReviewMutationVariables = {
    pid: string;
    commentBody: string;
};